fraud-monitoring/requirements.txt
apache-airflow[celery,postgres,redis,ssh,http]==2.8.1
apache-airflow-providers-amazon>=8.0.0
apache-airflow-providers-google>=10.0.0
apache-airflow-providers-microsoft-azure>=6.0.0
apache-airflow-providers-snowflake>=5.0.0
apache-airflow-providers-docker>=3.0.0
apache-airflow-providers-apache-spark>=4.0.0
apache-airflow-providers-apache-hive>=6.0.0
apache-airflow-providers-postgres
airflow-providers-clickhouse
boto3>=1.26.0
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=11.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
celery>=5.2.0
python-dotenv>=0.21.0
python-ldap>=3.4.0
cryptography>=38.0.0
kaggle>=1.8.0

